# Optimize Enumerable.OfType with custom iterator